requirements.txt
discord.py~=2.4.0
PyYAML~=6.0.2
python-dateutil~=2.9.0.post0
attrs~=24.2.0
jsonschema~=4.23.0
sqlmodel~=0.0.22
pandas~=2.2.3
plotly~=6.0.0
SQLAlchemy~=2.0.32
requests~=2.32.3
alembic~=1.13.2
aiohttp~=3.10.11
click~=8.1.7
uvicorn~=0.34.0
fastapi~=0.115.8
pydantic~=2.8.2
kaleido==0.2.1  # 0.1.0.post1 for windows